ServerList object now works correctly
Returns a single suitable node port in
json format. Previously, the reference
was lost due to incorrect structure.

The ServerList is now initialised inside
server.js, passed to routes.js and then
finally added as a reference to the req
object. This may in future be a global
variable array reference so we don't
need to define multiple middlewares
for every global we want to give to
routes.